// src/features/reportes/types/reporte-types.ts
export type MetricasAdministrativas = {
  total_estudiantes: number;
  total_cursos: number;
  total_inscripciones: number;
  ingresos: number;
  ganancias_instituto: number;
  ganancias_docentes: number;
  liquidaciones_pendientes: number;
  pagos_pendientes: number;
  pagos_realizados: number;
};